Reorganize pose CDK folders
- Move folders to suit UserCapture module organization
