+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42e9d85b4f00bb14f8ac84be70a6ed61
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
