+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects/PoseCondition/MagnitudePoseCondition.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b08c722f020796645950bdaf2611e31b
+guid: 701f5ae9d0f223e45957b4e4b74d565a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
