+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects/PoseCondition/ScalePoseCondition.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 598f6eee38b89ad4c85159e63c36b28a
+guid: a3b89854646245c4bb080daa9755a8fc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
