+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects/PoseConditionProcessor.cs.meta
