+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects/PoseOverriderContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6092cdef60cf0a345b8512bbc3d09583
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
