+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Objects/SkeletonPose.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d2cd4810b404cc04f8a9f45f863b2b9f
+guid: db8201a7302f194419a41a8a94b73106
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
